requirements.txt
python-dotenv
pinecone-client
langchain-community
langchain-text-splitters
langchain-pinecone
sentence-transformers
torch
